+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Reference docs: syntax highlighting for code samples
 - Navigate to _file:lineNumber_ location from [rg-asm](https://rg.atari.org/tools/rg-asm/) console messages
 - Support `rseven` directive
+- Editor inlay hint _CPU for Register_ (MC68010+ registers)
 
 ### Fixed